requirements.txt
yfinance==1.1.0
pandas
ta==0.11.0
PyYAML==6.0.2
schedule==1.2.2
pytz==2025.2
requests==2.32.3
python-dotenv==1.2.1
numpy
scikit-learn
openai
